+++ change_log/next/FE-1183.yml
@@ -1,1 +1,4 @@
-New Features: "Added functions to create a palette of colors."
+New Features: |
+Added a palette utility function which, given a configuration of
+base colors, will return functions to add tint or shade.
+
