+++ src/style/index.js
@@ -1,5 +1,5 @@
 import generatePalette from './palette';
-import atOpacity from './utils/at_opacity';
+import atOpacity from './utils/at-opacity';
 
 const blackAtOpacity = atOpacity('#FFFFFF');
 const addOpacity = (color, opacity) => atOpacity(color)(opacity);
