+++ src/style/themes/base/base-theme.config.js
