+++ src/style/themes/base/index.js
@@ -1,17 +1,21 @@
-import configureBase from './base_theme.config';
+import configureBase from './base-theme.config';
 import baseColors from '../../color-config';
-import atOpacity from '../../utils/at_opacity';
-import { mergeDeep } from '../../utils/merge_deep';
+import atOpacity from '../../utils/at-opacity';
+import { mergeDeep } from '../../utils/merge-deep';
 import generatePalette from '../../palette';
 
 const palette = {
   ...generatePalette(baseColors),
-  atOpacity
+  atOpacity,
+  baseColors
 };
 
-export default (configureTheme) => {
-  const baseTheme = configureBase(palette);
-  const themeToMergWithBase = configureTheme(palette);
+const baseTheme = configureBase(palette);
 
-  return mergeDeep(baseTheme, themeToMergWithBase);
+export default baseTheme;
+
+export const mergeWithBase = (configureTheme) => {
+  const themeToMergeWithBase = configureTheme(palette);
+
+  return mergeDeep(baseTheme, themeToMergeWithBase);
 };
