+++ src/style/themes/classic/classic-theme.config.js
