+++ src/style/themes/generic/generic-theme.config.js
