+++ src/style/themes/generic/index.js
@@ -1,4 +1,4 @@
-import configureTheme from './generic_theme.config';
-import mergeWithBase from '../base';
+import configureTheme from './generic-theme.config';
+import { mergeWithBase } from '../base';
 
 export default mergeWithBase(configureTheme);
