+++ src/style/themes/large-business/large-business-theme.config.js
