--- src/style/themes/large_business/index.js
@@ -1,4 +1,0 @@
-import configureTheme from './large_business_theme.config';
-import mergeWithBase from '../base';
-
-export default mergeWithBase(configureTheme);
