+++ src/style/themes/medium-business/medium-business-theme.config.js
