--- src/style/themes/medium_business/index.js
@@ -1,4 +1,0 @@
-import configureTheme from './medium_business_theme.config';
-import mergeWithBase from '../base';
-
-export default mergeWithBase(configureTheme);
