+++ src/style/themes/small-business/index.js
@@ -1,4 +1,4 @@
-import configureTheme from './classic-theme.config';
+import configureTheme from './small-business-theme.config';
 import { mergeWithBase } from '../base';
 
 export default mergeWithBase(configureTheme);
