+++ src/style/themes/small-business/small-business-theme.config.js
