--- src/style/themes/small_business/index.js
@@ -1,4 +1,0 @@
-import configureTheme from './small_business_theme.config';
-import mergeWithBase from '../base';
-
-export default mergeWithBase(configureTheme);
