+++ src/style/utils/merge-deep.js
